refactor: define a single parent table for all constriants
